feat: Clear grid on resize
Quite janky & hackish. Not the best way to do it but it works for now.
Biggest issue is the ghosting effect which is caused by transition & removing transition causes other glitches
Like the grid box from which the transition was removed makes it unresponsive to the hover effect.

Proper solution would be to first check if are there any filled boxes in the grid, if there are, first remove those & then return their count
Add or subtract this count to/from the newValue depending on if the grid is being created or removed

Sigh. One would argue that this optimization has led me down a path of complexity & it doesn't provide any benefits either.
Well, not significant benefits, the running time of our solution in the worse case is still O(n)