fix: removeEventListener logic to trigger on leaving canvas
This fixes the notorious bug when the mousemove event won't be removed if the user leaves the canvas
while still holding down the mouse button. It makes sense because mouseup event is registered on canvas
not on body, that's why the active tool would keep working.

This was fixed by adding a mouseleave event to the canvas, now, every time on leaving the canvas
it will trigger removeEventListener method & thus removing any mousemove events from the canvas