fix: NaN error on clicking /, %, and - operators
NaN appeared because in our multiple, divide & subtract functions we are
manually accessing individual elements of an array. And since we were calling
the operate function before checking if we have exactly 2 numbers or not.
It caused these 3 functions to execute (if the user chose them ofcourse)
with one number & the other thing which must be undefined. And that returned NaN

It didn't happen with reduce method, perhaps, because reduce stops looping when
it hits undefined.

We need to have exactly 2 numbers & an operator before calling operate.