fix: Only show = when the user presses it
Bugs:
Infinity zeroes in equation
Clear doesn't work properly when the user tries clearing immediately after
getting the answer, it populates the cleared number in equation

Also. this code is a huge mess. I couldn't have imagined that programming a calculator
can be so hard.

I can't decrease the complexity of it yet, I will come back to it after gaining more
experience and fix this shit