Library: Fix toggle button bug
toggling any other "book's" mark as read was only changing the state of the first
book on the page. The reason being input element IDs, since there can be only one
ID per page, the browser was looking through our page from top to bottom & when
it found the input element with an ID of "toggle" it immediately acted on it.

Fixed this issue by removing IDs, and since our input element is inside of Label
clicking on the label will still activate the input checkbox