TTT: Bugfix: Premature board clearing
Clicking on the board twice when a single cell is remaining on the board,
cleared the board.

This was happening because our model.insert method checks every cell against
the inserted value, and it was treating the board as a part of the cell.
Thus, it gets inserted & our set length becomes 9, stating that our board
was full when it was clearly not