TTT: Disable reset button when the result is being declared
Due to async & await functions, clicking on the reset game mess up the gameflow
Game gets reset but due to the away keyword, game progress normally.
Because resetGame is not async & the code which declares our result with a delay is.

A naive solution is to disable the reset button when a result is being
declared