KT: make generated moves keep a reference to their parent move
parent move refers to the move which caused the new move to be generated.
If [0, 0] is the parent, then [1, 2] & [2, 1] are its children. These children
keep track of their parent so that we can later follow this parent pointers
and form a "trail" or "path" of moves from the parent to the target